requirements.txt
flask
requests
ultralytics
pillow
opencv-python